# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Shop Liên Quân API - OpenAPI Spec",
    "version": "1.0.0",
    "description": "Tài liệu OpenAPI 3.0 / JSON API đầy đủ cho hệ thống website bán tài khoản game Liên Quân Mobile.",
    "contact": {
      "name": "API Support",
      "email": "support@shoplienquan.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT Access Token dạng: Bearer <token>"
      }
    },
...
  },
  "paths": {
    "/api/upload": {
      "post": {
        "summary": "Upload ảnh lên hệ thống",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File ảnh (jpg, png, webp...)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "url": {
                      "type": "string",
                      "example": "/uploads/1722400000000-image.png"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "summary": "Lấy lịch sử giao dịch của tôi",
        "tags": [
          "Transaction"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          }
        }
      }
    },
...
    "/api/profile": {
      "get": {
        "summary": "Lấy thông tin cá nhân",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin cá nhân thành công"
          },
          "401": {
            "description": "Chưa đăng nhập"
          }
        }
      }
    },
...
    "/api/ctv/accounts": {
      "get": {
        "summary": "Danh sách tài khoản do CTV đăng",
        "tags": [
          "CTV"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "0 đang bán, 1 đã bán, 2 đã ẩn"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách tài khoản của CTV thành công"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền CTV"
          }
        }
      }
    },
    "/api/ctv/orders": {
      "get": {
        "summary": "Danh sách đơn hàng từ tài khoản của CTV",
        "tags": [
          "CTV"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy đơn hàng của CTV thành công"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền CTV"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "summary": "Lấy danh sách danh mục tài khoản",
        "tags": [
          "Category"
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách danh mục thành công"
          }
        }
      },
      "post": {
        "summary": "Thêm danh mục tài khoản",
        "tags": [
          "Category"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Acc Liên Quân"
                  },
                  "noidung": {
                    "type": "string",
                    "example": "Danh mục bán tài khoản Liên Quân"
                  },
                  "type": {
                    "type": "string",
                    "example": "lienquan"
                  },
                  "status": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thêm danh mục thành công"
          },
          "401": {
            "description": "Vui lòng đăng nhập để tiếp tục"
          },
          "403": {
            "description": "Bạn không có quyền quản trị hệ thống"
          }
        }
      }
    },
...
    "/api/admin/accounts": {
      "get": {
        "summary": "Danh sách tài khoản game",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/orders": {
      "get": {
        "summary": "Danh sách đơn hàng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/transactions": {
      "get": {
        "summary": "Danh sách giao dịch",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/banks": {
      "get": {
        "summary": "Danh sách ngân hàng admin quản lý",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công"
          }
        }
      },
      "post": {
        "summary": "Thêm ngân hàng nạp tiền mới",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bank_name",
                  "account_number",
                  "account_holder"
                ],
                "properties": {
                  "bank_name": {
                    "type": "string",
                    "example": "MBBank"
                  },
                  "account_number": {
                    "type": "string",
                    "example": "0987654321"
                  },
                  "account_holder": {
                    "type": "string",
                    "example": "NGUYEN VAN A"
                  },
                  "qr_template": {
                    "type": "string",
                    "example": "https://img.vietqr.io/image/MB-0987654321-compact2.png"
                  },
                  "status": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thêm thành công"
          }
        }
      }
    },
...
    "/api/account-types": {
      "get": {
        "summary": "Lấy danh sách loại tài khoản",
        "tags": [
          "AccountType"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "danhmuc_id",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "Lọc theo ID danh mục"
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách loại tài khoản thành công"
          }
        }
      },
      "post": {
        "summary": "Thêm loại tài khoản",
        "tags": [
          "AccountType"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "danhmuc_id",
                  "name"
                ],
                "properties": {
                  "danhmuc_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Nick rank cao"
                  },
                  "img": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                  },
                  "noidung": {
                    "type": "string",
                    "example": "Loại tài khoản có rank cao, nhiều tướng"
                  },
                  "camket": {
                    "type": "string",
                    "example": "Cam kết đúng thông tin, bảo hành theo chính sách shop"
                  },
                  "status": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thêm loại tài khoản thành công"
          },
          "401": {
            "description": "Vui lòng đăng nhập để tiếp tục"
          },
          "403": {
            "description": "Bạn không có quyền quản trị hệ thống"
          }
        }
      }
    },
...
    "/api/accounts": {
      "get": {
        "summary": "Lấy danh sách tài khoản đang bán",
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "loai_id",
            "schema": {
              "type": "integer"
            },
            "description": "Lọc theo loại tài khoản"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "price_asc",
                "price_desc"
              ]
            },
            "description": "Sắp xếp theo giá"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 20
            }
          }
        ]
      },
      "post": {
        "summary": "Thêm tài khoản game",
        "tags": [
          "Account"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "loai_id",
                  "gia",
                  "login"
                ],
                "properties": {
                  "loai_id": {
                    "type": "integer"
                  },
                  "thong_tin": {
                    "type": "string"
                  },
                  "list_thong_tin": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string"
                  },
                  "list_img": {
                    "type": "string"
                  },
                  "login": {
                    "type": "string"
                  },
                  "gia": {
                    "type": "integer"
                  },
                  "ck": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}